+++ CP/core/set_domain.h
@@ -10,16 +10,16 @@
 class Set_Domain: public Domain    /// This class implements enumerated domains with event management \ingroup core
 {
 	protected :
-		vector<int> real_value;  		 					  ///< maps each value to its real value (useful for some predicate or global constraints)
-		map<int,int> real_value_to_value;			  ///< maps each real value to its value (useful for some global constraints)
+		vector<long> real_value;  		 					  ///< maps each value to its real value (useful for some predicate or global constraints)
+		map<long,int> real_value_to_value;			  ///< maps each real value to its value (useful for some global constraints)
 
 	public :
 		// constructor
-    Set_Domain (Event_Manager * evm, unsigned int num_var, set<int> & value_set);   		///< constructs a domain defined by the set value_set
+    Set_Domain (Event_Manager * evm, unsigned int num_var, set<long> & value_set);   		///< constructs a domain defined by the set value_set
 		
 		// basic functions	
-    int Get_Value (int real_val) override;	  ///< returns the value corresponding to the real value real_val if it exists, -1 otherwise
-    int Get_Real_Value (int num) override;		///< returns the real value corresponding to the value num
+    int Get_Value (long real_val) override;	  ///< returns the value corresponding to the real value real_val if it exists, -1 otherwise
+    long Get_Real_Value (int num) override;		///< returns the real value corresponding to the value num
 };
 
 
@@ -27,11 +27,11 @@ class Set_Domain: public Domain    /// This class implements enumerated domains 
 // inline function definitions
 //-----------------------------
 
-inline int Set_Domain::Get_Value (int real_val)
+inline int Set_Domain::Get_Value (long real_val)
 // returns the value corresponding to the real value real_val if it exists, -1 otherwise
 /// \param[in] real_val the real value for which we want to get the corresponding value in the domain
 {
-	map<int,int>::iterator iter = real_value_to_value.find (real_val);
+	map<long,int>::iterator iter = real_value_to_value.find (real_val);
 	if (iter == real_value_to_value.end())
 		return -1;
 	else return iter->second;
@@ -39,7 +39,7 @@ inline int Set_Domain::Get_Value (int real_val)
 }
 
 
-inline int Set_Domain::Get_Real_Value (int num)
+inline long Set_Domain::Get_Real_Value (int num)
 // returns the real value corresponding to the value num
 /// \param[in] num the number of the wished real value
 {
